pyproject.toml
[project]
name = "currenyalert"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.33.1",
]
